refactor InfoCard to use Card as value and moved Chart into Totals scope
